requirements.txt

boto3==1.34.49
botocore==1.34.49
ffmpeg-python==0.2.0
streamlit==1.37.0
streamlit-option-menu==0.3.13
streamlit-tree-select==0.0.5
python-dotenv==1.0.1